Moved end-host spkt processing to a different package
