+++ go/lib/spkt/spkt.go
@@ -36,89 +36,6 @@ type ScnPkt struct {
 	Pld     common.Payload
 }
 
-func ScnPktFromRaw(b common.RawBytes) (*ScnPkt, error) {
-	s := &ScnPkt{}
-	if err := s.Parse(b); err != nil {
-		return nil, err
-	}
-	return s, nil
-}
-
-func NewScnPkt() *ScnPkt {
-	return &ScnPkt{
-		CmnHdr: &CmnHdr{},
-		DstIA:  &addr.ISD_AS{},
-		SrcIA:  &addr.ISD_AS{},
-		Path:   &spath.Path{},
-		// Rest of fields passed by reference
-	}
-}
-
-func (s *ScnPkt) Parse(b common.RawBytes) error {
-	var cerr *common.Error
-	offset := 0
-
-	// Parse common header
-	if cerr = s.CmnHdr.Parse(b[:CmnHdrLen]); cerr != nil {
-		return cerr
-	}
-	offset += CmnHdrLen
-
-	// Parse address header
-	s.DstIA.Parse(b[offset:])
-	offset += addr.IABytes
-	s.SrcIA.Parse(b[offset:])
-	offset += addr.IABytes
-	if s.DstHost, cerr = addr.HostFromRaw(b[offset:], s.CmnHdr.DstType); cerr != nil {
-		return common.NewError("Unable to parse destination host address",
-			"err", cerr)
-	}
-	offset += s.DstHost.Size()
-	if s.SrcHost, cerr = addr.HostFromRaw(b[offset:], s.CmnHdr.SrcType); cerr != nil {
-		return common.NewError("Unable to parse source host address",
-			"err", cerr)
-	}
-	offset += s.SrcHost.Size()
-
-	// Validate address padding bytes
-	addrHdrLen := 2*addr.IABytes + s.DstHost.Size() + s.SrcHost.Size()
-	padBytes := util.CalcPadding(addrHdrLen, common.LineLen)
-	if pos, ok := isZeroMemory(b[offset : offset+padBytes]); !ok {
-		return common.NewError("Invalid padding", "position", pos,
-			"expected", 0, "actual", b[offset+pos])
-	}
-	offset += padBytes
-	addrHdrLen += padBytes
-
-	// Parse path header
-	pathLen := s.CmnHdr.HdrLenBytes() - CmnHdrLen - addrHdrLen
-	s.Path.Raw = b[offset : offset+pathLen]
-	s.Path.InfOff = s.CmnHdr.InfoFOffBytes()
-	s.Path.HopOff = s.CmnHdr.HopFOffBytes()
-	offset += pathLen
-
-	// TODO(scrye): Add extension support
-
-	// Parse L4 header
-	if s.CmnHdr.NextHdr != common.L4UDP {
-		return common.NewError("Unsupported NextHdr value", "expected",
-			common.L4UDP, "actual", s.CmnHdr.NextHdr)
-	}
-	if s.L4, cerr = l4.UDPFromRaw(b[offset : offset+l4.UDPLen]); cerr != nil {
-		return common.NewError("Unable to parse UDP header", "err", cerr)
-	}
-	offset += s.L4.L4Len()
-
-	// Parse payload
-	pldLen := int(s.CmnHdr.TotalLen) - s.CmnHdr.HdrLenBytes() - s.L4.L4Len()
-	if offset+pldLen < len(b) {
-		return common.NewError("Incomplete packet, bad payload length",
-			"expected", pldLen, "actual", len(b)-offset)
-	}
-	s.Pld = common.RawBytes(b[offset:])
-	return nil
-}
-
 func (s *ScnPkt) Copy() *ScnPkt {
 	c := &ScnPkt{}
 	if s.DstIA != nil {
@@ -193,13 +110,4 @@ func (s *ScnPkt) TotalLen() int {
 		l += s.Pld.Len()
 	}
 	return l
-}
-
-func isZeroMemory(b common.RawBytes) (int, bool) {
-	for i := range b {
-		if b[i] != 0 {
-			return i, false
-		}
-	}
-	return 0, true
 }
